preprocessing/pyproject.toml
[project]
name = "preprocessing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets==2.14.6",
    "fastapi==0.104.1",
    "google-cloud-storage==2.10.0",
    "pandas==2.1.3",
    "pydantic==2.5.0",
    "python-multipart==0.0.6",
    "uvicorn[standard]==0.24.0",
    "werkzeug==3.0.1",
]
